+++ ui/src/components/General/Notification/utils/types.ts
@@ -10,9 +10,11 @@ interface NotificationPropsType {
 
 interface NotificationListPropsType {
 	notifications: Markdown[] | undefined;
+	showRedFlag: boolean;
 	colorText: string;
 	isLoading: boolean;
 	isError: boolean;
+	handleRedFlagNotification: () => void;
 }
 
 export type { Markdown, NotificationPropsType, NotificationListPropsType };
