+++ ui/src/components/Layouts/FloatingHeader/index.tsx
@@ -1,7 +1,7 @@
 import { Avatar, Dropdown, Space, theme } from "antd";
 import style from "./FloatingHeader.module.scss";
-import { items } from "./utils/constants";
-import Notification from "./components/Notification";
+import { items } from "./utils/contants";
+import Notification from "components/General/Notification";
 import avatar1 from "assets/avatars/avatar-1.svg";
 import avatar2 from "assets/avatars/avatar-2.svg";
 import avatar3 from "assets/avatars/avatar-3.svg";
