+++ ui/src/pages/Markdown/MarkdownEditor/MarkdownEditor.module.scss
@@ -7,5 +7,6 @@
 	&__editor {
 		flex: 1;
 		margin-top: 20px;
+		overflow: auto;
 	}
 }
