+++ ui/src/pages/Markdown/MdTableGenerator/MdTableGenerator.module.scss
@@ -6,5 +6,6 @@
 
 	&__editor {
 		flex: 1;
+		overflow: auto;
 	}
 }
