+++ ui/src/pages/Markdown/TableOfContent/tests/TableOfContent.test.tsx
@@ -1,0 +1,9 @@
+import { render } from "@testing-library/react";
+import { TableOfContent } from "pages/pages";
+import { describe, test } from "vitest";
+
+describe("Table Of content", () => {
+	test("Render component without crash", () => {
+		render(<TableOfContent />);
+	});
+});
